Move functions to different files
